+++ axum-macros/src/typed_path.rs
@@ -29,7 +29,7 @@ pub(crate) fn expand(item_struct: ItemStruct) -> syn::Result<TokenStream> {
             let segments = parse_path(&path)?;
             expand_unnamed_fields(fields, ident, path, &segments)
         }
-        syn::Fields::Unit => Ok(expand_unit_fields(ident, path)?),
+        syn::Fields::Unit => expand_unit_fields(ident, path),
     }
 }