+++ axum-extra/CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-- None.
+- **added:** Add `TypedPath::to_uri` for converting the path into a `Uri` ([#790])
+
+[#790]: https://github.com/tokio-rs/axum/pull/790
 
 # 0.1.4 (22. February, 2022)